:recycle: Redis 설정 리팩터링
리팩터링
