+++ build.gradle
@@ -30,6 +30,7 @@ dependencies {
 
 	//redis
 	implementation 'org.springframework.boot:spring-boot-starter-data-redis'
+	implementation 'org.springframework.session:spring-session-data-redis'
 
 	//mail
 	implementation 'org.springframework.boot:spring-boot-starter-mail'
