--- src/main/java/msg/team1/Hi/global/common/RedisDao.java
@@ -1,35 +1,0 @@
-package msg.team1.Hi.global.common;
-
-import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.stereotype.Component;
-
-import java.time.Duration;
-
-@Component
-public class RedisDao {
-    private final RedisTemplate<String, String> redisTemplate;
-
-    public RedisDao(RedisTemplate<String, String> redisTemplate) {
-        this.redisTemplate = redisTemplate;
-    }
-
-    public void setValues(String key, String data) {
-        ValueOperations<String, String> values = redisTemplate.opsForValue();
-        values.set(key, data);
-    }
-
-    public void setValues(String key, String data, Duration duration) {
-        ValueOperations<String, String> values = redisTemplate.opsForValue();
-        values.set(key, data, duration);
-    }
-
-    public String getValues(String key) {
-        ValueOperations<String, String> values = redisTemplate.opsForValue();
-        return values.get(key);
-    }
-
-    public void deleteValues(String key) {
-        redisTemplate.delete(key);
-    }
-}
