+++ src/main/java/msg/team1/Hi/global/redis/properties/RedisProperties.java
@@ -1,0 +1,15 @@
+package msg.team1.Hi.global.redis.properties;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+import org.springframework.boot.context.properties.ConstructorBinding;
+
+@Getter
+@ConstructorBinding
+@RequiredArgsConstructor
+@ConfigurationProperties(prefix = "spring.redis")
+public class RedisProperties {
+    private final String host;
+    private final Integer port;
+}
