--- src/main/java/msg/team1/Hi/global/util/RedisUtil.java
@@ -1,39 +1,0 @@
-package msg.team1.Hi.global.util;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.redis.core.StringRedisTemplate;
-import org.springframework.data.redis.core.ValueOperations;
-import org.springframework.stereotype.Service;
-
-import java.time.Duration;
-
-@Service
-@RequiredArgsConstructor
-public class RedisUtil {
-
-
-    private final StringRedisTemplate redisTemplate;
-
-    // key 를 통해 value 리턴
-    public String getData(String key) {
-        ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
-        return valueOperations.get(key);
-    }
-
-    public void setData(String key, String value) {
-        ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
-        valueOperations.set(key, value);
-    }
-
-    // 유효 시간 동안 (key, value) 저장
-    public void setDataExpire(String key, String value, long duration) {
-        ValueOperations<String, String> valueOperations = redisTemplate.opsForValue();
-        Duration expireDuration = Duration.ofSeconds(duration);
-        valueOperations.set(key, value, expireDuration);
-    }
-
-    // 삭제
-    public void deleteData(String key) {
-        redisTemplate.delete(key);
-    }
-}
